# openapi/LupiraMtgApi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Lupira MTG API",
    "description": "Backend for the Lupira MTG mobile app — Magic: The Gathering card metadata, scan-based recognition, and per-user collection management. Authenticate with a Bearer token issued by Authentik (OIDC).",
    "version": "v1"
  },
  "paths": {
...
    "/admin/set-type-weights": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List set-type weights used by the scan-confidence ranker.",
        "description": "The scan ranker biases candidates by their printing's set type — e.g. weighting\ncore/expansion above commander/funny so a Lightning Bolt scan prefers the M21\nprinting over a Funko Pop one. This endpoint exposes the live values.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetTypeWeightListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/admin/set-type-weights/{setType}": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Upsert the weight for one set type.",
        "description": "Body: `{ weight: number }`. Weight must be finite and non-negative; 1.0 is the\nneutral baseline, < 1 down-weights, > 1 up-weights. New `setType` keys are created\non the fly so you don't have to pre-seed.\n\nThe change takes effect on the *next* scan; the ranker reads weights per request.",
        "parameters": [
          {
            "name": "setType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSetTypeWeightRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetTypeWeightResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/sets": {
      "get": {
        "tags": [
          "Sets"
        ],
        "summary": "Browse the catalog of Magic sets.",
        "description": "Returns paginated set metadata. Use this to power a set picker (the `?set=`\nfilter on `GET /cards` takes the same `code` returned here).\n\n`setType` filters by Scryfall set type (`core`, `expansion`, `masters`,\n`commander`, `funny`, …).\n`sort` ∈ `releasedAt|code|name`, default `releasedAt`.\n`order` ∈ `asc|desc`, default `desc` (newest first).\n`take` 1–200 (default 50). `skip` for paging.",
        "parameters": [
          {
            "name": "setType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/sets/{code}": {
      "get": {
        "tags": [
          "Sets"
        ],
        "summary": "Get one set by its lower-case code.",
        "description": "Returns set metadata plus a presigned URL for the set icon (if cached locally).",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SetListResponse": {
        "required": [
          "results",
          "total"
        ],
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SetResponse"
            }
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "SetResponse": {
        "required": [
          "code",
          "name",
          "setType"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "setType": {
            "type": "string"
          },
          "releasedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "cardCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "iconUrl": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "SetTypeWeightListResponse": {
        "required": [
          "weights"
        ],
        "type": "object",
        "properties": {
          "weights": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SetTypeWeightResponse"
            }
          }
        }
      },
      "SetTypeWeightResponse": {
        "required": [
          "setType",
          "weight",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "setType": {
            "type": "string"
          },
          "weight": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UpdateSetTypeWeightRequest": {
        "required": [
          "weight"
        ],
        "type": "object",
        "properties": {
          "weight": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "PoC device token from `POST /me/register`. Send as `Authorization: Bearer lmtg_<token>`. Replaced by Authentik OIDC tokens in the future Path C migration (see plan).",
        "scheme": "bearer"
      }
    }
  },
...
}